Create main.cpp
This is the bare shell code structure and the starting line for the main.cpp
Familiarize yourself with the structure hence you start to understand and relate which program logic will belong to which section. 
We have two sections to begin with. 
Setup - program logic that stages the esp32 
Loop -  Program logic that runs repeatedly (can be for inputting, receiving, outputting, transmitting data, continuously )

Once we understand this logic. We can start our journey exploring microcontrollers..